installer fixes
